add: separate TopBar that uses GDStore and causes wallet/userstorage loading
