+++ src/components/auth/Auth.js
@@ -4,7 +4,8 @@ import normalize from 'react-native-elements/src/helpers/normalizeText'
 import { View } from 'react-native'
 import Mnemonics from '../signin/Mnemonics'
 import logger from '../../lib/logger/pino-logger'
-import { CustomButton, Text } from '../common'
+import Text from '../common/view/Text'
+import CustomButton from '../common/buttons/CustomButton'
 import { Description, LinkButton, Title } from '../signup/components'
 import { PrivacyPolicy, TermsOfUse } from '../webView/webViewInstances'
 import { createStackNavigator } from '../appNavigation/stackNavigation'
