+++ src/components/dashboard/Claim.js
@@ -11,7 +11,8 @@ import SimpleStore from '../../lib/undux/SimpleStore'
 import { useDialog } from '../../lib/undux/utils/dialog'
 import wrapper from '../../lib/undux/utils/wrapper'
 import { weiToMask } from '../../lib/wallet/utils'
-import { CustomButton, Section, Text, TopBar, Wrapper } from '../common'
+import { CustomButton, Section, Text, Wrapper } from '../common'
+import TopBar from '../common/view/TopBar'
 import ErrorIcon from '../common/modal/ErrorIcon'
 import LoadingIcon from '../common/modal/LoadingIcon'
 import SuccessIcon from '../common/modal/SuccessIcon'
