+++ src/components/dashboard/Send.js
@@ -8,7 +8,8 @@ import { useDialog } from '../../lib/undux/utils/dialog'
 import InputRecipient from '../common/form/InputRecipient'
 import isMobilePhone from '../../lib/validators/isMobilePhone'
 import goodWallet from '../../lib/wallet/GoodWallet'
-import { CustomButton, IconButton, Section, TopBar, Wrapper } from '../common'
+import { CustomButton, IconButton, Section, Wrapper } from '../common'
+import TopBar from '../common/view/TopBar'
 import { routeAndPathForCode } from './utils/routeAndPathForCode'
 
 const SEND_TITLE = 'Send G$'
