+++ src/components/dashboard/SendConfirmation.web.js
@@ -6,7 +6,8 @@ import { isMobile } from 'mobile-device-detect'
 import { generateSendShareObject } from '../../lib/share'
 import { useDialog } from '../../lib/undux/utils/dialog'
 import { useScreenState } from '../appNavigation/stackNavigation'
-import { BigGoodDollar, CopyButton, CustomButton, QRCode, Section, Text, TopBar, Wrapper } from '../common'
+import { BigGoodDollar, CopyButton, CustomButton, QRCode, Section, Text, Wrapper } from '../common'
+import TopBar from '../common/view/TopBar'
 import DoneButton from '../common/buttons/DoneButton'
 import { withStyles } from '../../lib/styles'
 import { getScreenHeight } from '../../lib/utils/Orientation'
