+++ src/components/signup/SmsForm.web.js
@@ -5,7 +5,8 @@ import OtpInput from 'react-otp-input'
 import normalize from 'react-native-elements/src/helpers/normalizeText'
 import logger from '../../lib/logger/pino-logger'
 import API from '../../lib/API/api'
-import { LoadingIndicator, Text } from '../common'
+import Text from '../common/view/Text'
+import LoadingIndicator from '../common/view/LoadingIndicator'
 import { ActionButton, Description, Error, Title, Wrapper } from './components'
 import type { SignupState } from './SignupState'
