+++ src/components/splash/Splash.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { Image, StyleSheet, View } from 'react-native'
 import normalize from 'react-native-elements/src/helpers/normalizeText'
 import splashImage from '../../assets/splash.png'
-import { Text } from '../common'
+import Text from '../common/view/Text'
 
 //minimize delay <Image> has over web <img>
 Image.prefetch(splashImage)
